# api-docs.yml
...
openapi: 3.0.3
info:
  title: UltraZend SMTP Server API
  description: |
    # 🚀 UltraZend SMTP Server API
    
    **Servidor SMTP profissional com API REST completa para envio e gerenciamento de emails.**
    
    ## 🌟 Funcionalidades
    
    - **Envio de emails** via API REST
    - **Servidor SMTP** completo (MX + Submission)  
    - **Autenticação JWT** com refresh tokens
    - **Rate limiting** inteligente
    - **Webhooks** para notificações
    - **Analytics** em tempo real
    - **Sistema de filas** Redis Bull
    - **Monitoramento** Prometheus
    - **Segurança** multicamadas
    
    ## 🔐 Autenticação
    
    A API utiliza **JWT Bearer tokens** para autenticação:
    
    ```bash
    curl -H "Authorization: Bearer YOUR_JWT_TOKEN" \
         -H "Content-Type: application/json" \
         https://api.ultrazend.com.br/api/emails/send
    ```
    
    ## 🚦 Rate Limiting
    
    - **API**: 1000 requests/hora por usuário
    - **Email Send**: 10,000 emails/hora por usuário
    - **Auth**: 10 tentativas/15min por IP
    
    ## 📧 SMTP Configuration
    
    **MX Server (recebimento):**
    - Host: `mail.ultrazend.com.br`
    - Port: `25`
    
    **Submission Server (envio autenticado):**
    - Host: `mail.ultrazend.com.br`
    - Port: `587`
    - Security: `STARTTLS`
    
    ## 🔗 Links Úteis
    
    - **Dashboard**: https://www.ultrazend.com.br
    - **Status Page**: https://status.ultrazend.com.br
    - **Support**: suporte@ultrazend.com.br
...
    
  version: 2.0.0
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtido através do endpoint de login.
        
        Exemplo: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API Key para integração direta sem login.
        
        Exemplo: `X-API-Key: ultrazend_ak_1234567890abcdef`
  schemas:
...
    Email:
      type: object
      properties:
        id:
          type: integer
          example: 456
        from:
          type: string
          format: email
          example: "sender@ultrazend.com.br"
        to:
          type: string
          format: email
          example: "recipient@example.com"
        subject:
          type: string
          example: "Bem-vindo ao UltraZend!"
        status:
          type: string
          enum: [queued, processing, delivered, failed, bounced]
          example: "delivered"
        created_at:
          type: string
          format: date-time
          example: "2025-01-01T12:00:00.000Z"
        delivered_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-01-01T12:01:30.000Z"
        message_id:
          type: string
          example: "msg-abc123def456"
        error_message:
          type: string
          nullable: true
          example: null
      required:
        - id
        - from
        - to
        - subject
        - status
    EmailSendRequest:
      type: object
      properties:
        from:
          type: string
          format: email
          example: "noreply@ultrazend.com.br"
          description: "Endereço do remetente (deve estar autorizado)"
        to:
          type: string
          format: email
          example: "cliente@example.com"
          description: "Endereço do destinatário"
        subject:
          type: string
          maxLength: 200
          example: "Bem-vindo ao nosso serviço!"
          description: "Assunto do email"
        html:
          type: string
          example: "<h1>Olá!</h1><p>Bem-vindo ao UltraZend!</p>"
          description: "Conteúdo HTML do email"
        text:
          type: string
          example: "Olá! Bem-vindo ao UltraZend!"
          description: "Conteúdo em texto puro (fallback)"
        reply_to:
          type: string
          format: email
          example: "contato@ultrazend.com.br"
          description: "Endereço para respostas"
        cc:
          type: array
          items:
            type: string
            format: email
          example: ["cc@example.com"]
          description: "Destinatários em cópia"
        bcc:
          type: array
          items:
            type: string
            format: email
          example: ["bcc@example.com"]
          description: "Destinatários em cópia oculta"
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/EmailAttachment'
          description: "Anexos do email"
        template_id:
          type: string
          example: "welcome-template"
          description: "ID do template a ser usado"
        template_vars:
          type: object
          example: {"name": "João", "company": "ACME Inc"}
          description: "Variáveis para o template"
        tags:
          type: array
          items:
            type: string
          example: ["newsletter", "welcome"]
          description: "Tags para categorização"
        metadata:
          type: object
          example: {"campaign_id": "camp123", "source": "website"}
          description: "Metadados personalizados"
        scheduled_at:
          type: string
          format: date-time
          example: "2025-01-02T09:00:00.000Z"
          description: "Agendar envio para data/hora específica"
        webhook_url:
          type: string
          format: uri
          example: "https://yourapp.com/webhook/email-status"
          description: "URL para receber notificações de status"
      required:
        - from
        - to
        - subject
      oneOf:
        - required: [html]
        - required: [text]
        - required: [template_id]
    EmailAttachment:
      type: object
      properties:
        filename:
          type: string
          example: "documento.pdf"
        content:
          type: string
          format: base64
          example: "JVBERi0xLjQK..."
        content_type:
          type: string
          example: "application/pdf"
        size:
          type: integer
          example: 1024000
          description: "Tamanho em bytes (máx: 25MB)"
      required:
        - filename
        - content
        - content_type

    # === RESPONSE SCHEMAS ===
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operação realizada com sucesso"
        data:
          type: object
          description: "Dados específicos da operação"
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Erro na validação dos dados"
        code:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: object
          description: "Detalhes específicos do erro"
        trace_id:
          type: string
          example: "req-abc123def456"
      required:
        - success
        - error
        - code

    # === HEALTH CHECK SCHEMAS ===
...
  responses:
    UnauthorizedError:
      description: Token de acesso inválido ou expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Token de acesso inválido"
            code: "UNAUTHORIZED"
            trace_id: "req-abc123def456"
    ValidationError:
      description: Erro de validação nos dados enviados
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Dados inválidos"
            code: "VALIDATION_ERROR"
            details:
              email: ["Email é obrigatório"]
              password: ["Senha deve ter pelo menos 8 caracteres"]
            trace_id: "req-abc123def456"
    RateLimitError:
      description: Limite de taxa excedido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Limite de taxa excedido"
            code: "RATE_LIMIT_EXCEEDED"
            details:
              limit: 1000
              window: "1 hour"
              retry_after: 3600
            trace_id: "req-abc123def456"
...
paths:
...
  /emails/send:
    post:
      tags:
        - Email Sending
      summary: Enviar email único
      description: |
        Envia um email para um destinatário específico.
        
        **Recursos suportados:**
        - Templates dinâmicos
        - Anexos (até 25MB total)
        - Agendamento
        - Webhooks de status
        - Tags e metadados
        - HTML + Text
        
        **Rate Limits:**
        - 10,000 emails/hora (plano Pro)
        - 1,000 emails/hora (plano Free)
      operationId: sendEmail
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailSendRequest'
            examples:
              simple_email:
                summary: Email simples
                value:
                  from: "noreply@ultrazend.com.br"
                  to: "cliente@example.com"
                  subject: "Bem-vindo!"
                  html: "<h1>Olá!</h1><p>Bem-vindo ao nosso serviço.</p>"
                  text: "Olá! Bem-vindo ao nosso serviço."
              
              template_email:
                summary: Email com template
                value:
                  from: "noreply@ultrazend.com.br"
                  to: "cliente@example.com"
                  subject: "Bem-vindo, {{name}}!"
                  template_id: "welcome-template"
                  template_vars:
                    name: "João"
                    company: "ACME Inc"
                  tags: ["welcome", "onboarding"]
                  webhook_url: "https://myapp.com/webhook/email-status"
      responses:
        '202':
          description: Email aceito para processamento
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          email_id:
                            type: string
                            example: "email-abc123def456"
                          message_id:
                            type: string
                            example: "msg-789xyz"
                          status:
                            type: string
                            example: "queued"
                          scheduled_for:
                            type: string
                            format: date-time
                            nullable: true
                            example: null
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'

  # === GERENCIAMENTO DE EMAILS ===
  /emails:
    get:
      tags:
        - Email Management
      summary: Listar emails enviados
      description: |
        Lista emails enviados pelo usuário com opções de filtro e ordenação.
      operationId: listEmails
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Número da página
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Itens por página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: status
          in: query
          description: Filtrar por status
          schema:
            type: string
            enum: [queued, processing, delivered, failed, bounced]
      responses:
        '200':
          description: Lista de emails
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Email'
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      per_page:
                        type: integer
                        example: 25
                      total:
                        type: integer
                        example: 250
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /emails/{emailId}:
    get:
      tags:
        - Email Management
      summary: Obter detalhes do email
      description: Retorna informações detalhadas de um email específico
      operationId: getEmail
      security:
        - bearerAuth: []
      parameters:
        - name: emailId
          in: path
          required: true
          description: ID do email
          schema:
            type: string
            example: "email-abc123def456"
      responses:
        '200':
          description: Detalhes do email
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Email'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Email não encontrado

# ========================================
# TAGS PARA ORGANIZAÇÃO
# ========================================
...